Feat: adicionando autenticação JWT nas operações de produtos

+++ src/produtos/produtos.service.ts
@@ -7,10 +7,14 @@ import { UpdateProdutoDto } from 'src/dto/produtos/update-produto.dto';
 export class ProdutosService {
   constructor(private prisma: PrismaService) {}
 
-  async getAll() {
-    const produtos = await this.prisma.produto.findMany();
+  async getAll(userId: number) {
+    const produtos = await this.prisma.produto.findMany({
+      where: { usuarioId: userId },
+    });
     if (!produtos.length) {
-      throw new NotFoundException('Nenhum produto encontrado.');
+      throw new NotFoundException(
+        'Nenhum produto encontrado para este usuário.',
+      );
     }
     return produtos;
   }
@@ -25,10 +29,13 @@ export class ProdutosService {
     return produto;
   }
 
-  async create(produto: CreateProdutoDto) {
+  async create(produto: CreateProdutoDto, userId: number) {
     try {
       const novoProduto = await this.prisma.produto.create({
-        data: produto,
+        data: {
+          ...produto,
+          usuario: { connect: { id: userId } },
+        },
       });
       return novoProduto;
     } catch (error) {
@@ -36,11 +43,14 @@ export class ProdutosService {
     }
   }
 
-  async update(id: number, produto: UpdateProdutoDto) {
+  async update(id: number, produto: UpdateProdutoDto, userId: number) {
     try {
       const produtoAtualizado = await this.prisma.produto.update({
-        where: { id },
-        data: produto,
+        where: { id, usuarioId: userId },
+        data: {
+          ...produto,
+          usuario: { connect: { id: userId } },
+        },
       });
       return produtoAtualizado;
     } catch (error) {
@@ -49,15 +59,30 @@ export class ProdutosService {
     }
   }
 
-  async delete(id: number) {
+  async delete(id: number, userId: number) {
     try {
-      const produtoDeletado = await this.prisma.produto.delete({
-        where: { id },
+      const produto = await this.prisma.produto.findUnique({
+        where: { id, usuarioId: userId },
       });
+
+      if (!produto || produto.usuarioId !== userId) {
+        throw new NotFoundException(
+          `Produto com ID ${id} não encontrado ou não pertence ao usuário.`,
+        );
+      }
+
+      const produtoDeletado = await this.prisma.produto.delete({
+        where: {
+          id,
+          usuarioId: userId,
+        },
+      });
+
       return produtoDeletado;
     } catch (error) {
-      throw new Error(`
-        Erro ao deletar o produto com ID ${id}: ${error.message}`);
+      throw new Error(
+        `Erro ao deletar o produto com ID ${id}: ${error.message}`,
+      );
     }
   }
 }
